requirements.txt
instaloader==4.10
PySocks==1.7.1
google-api-python-client==2.103.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
neo4j==5.13.0
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
